refactor(sidebar): split the tree coordinator and derive its shape once
