+++ TablePro/Views/Sidebar/DatabaseTreeOutlineCoordinator.swift
@@ -9,46 +9,48 @@ import TableProPluginKit
 
 @MainActor
 final class DatabaseTreeOutlineCoordinator: NSObject {
-    private weak var outlineView: NSOutlineView?
-    private let service = DatabaseTreeMetadataService.shared
+    internal weak var outlineView: NSOutlineView?
+    internal let service = DatabaseTreeMetadataService.shared
     private static let cellIdentifier = NSUserInterfaceItemIdentifier("DatabaseTreeCell")
 
-    private var connectionId = UUID()
-    private var databaseType: DatabaseType = .mysql
-    private weak var mainCoordinator: MainContentCoordinator?
-    private var windowState: WindowSidebarState?
-    private var sidebarState: SharedSidebarState?
-    private weak var viewModel: SidebarViewModel?
-    private var searchText = ""
+    internal var connectionId = UUID()
+    internal var databaseType: DatabaseType = .mysql
+    internal weak var mainCoordinator: MainContentCoordinator?
+    internal var windowState: WindowSidebarState?
+    internal var sidebarState: SharedSidebarState?
+    internal weak var viewModel: SidebarViewModel?
+    internal var searchText = ""
     private var isConnected = false
-    private var activeDatabase: String?
+    internal var activeDatabase: String?
     private var activeSchema: String?
     private var pendingTruncates: Set<String> = []
     private var pendingDeletes: Set<String> = []
-    private var showRecentTables = true
+    internal var showRecentTables = true
 
-    private var nodeCache: [String: DatabaseTreeNode] = [:]
-    private var childrenCache: [String: [DatabaseTreeNode]] = [:]
+    internal var nodeCache: [String: DatabaseTreeNode] = [:]
+    internal var childrenCache: [String: [DatabaseTreeNode]] = [:]
+    private var cachedRowContext: DatabaseTreeRowContext?
+    private var cachedRowActions: DatabaseTreeRowActions?
     private var lastSelection: Set<DatabaseTreeTableRef> = []
     private var lastSelectedNodeIds: [String] = []
     private var publishedTables: Set<TableInfo> = []
     private var pendingOpenWork: DispatchWorkItem?
-    private var isApplyingExpansion = false
+    internal var isApplyingExpansion = false
     private var isSyncingSelection = false
     private var isReloading = false
     private var hasRenderedOnce = false
     private var reconcileScheduled = false
     private var observationGeneration = 0
 
-    private let schemaService = SchemaService.shared
+    internal let schemaService = SchemaService.shared
     private var favoriteTables: Set<FavoriteTablesStorage.FavoriteEntry> = []
     private var favoritesObserver: (any NSObjectProtocol)?
 
-    private var supportsSchemaLevel: Bool {
+    internal var supportsSchemaLevel: Bool {
         PluginManager.shared.databaseGroupingStrategy(for: databaseType) == .bySchema
     }
 
-    private var systemSchemas: Set<String> {
+    internal var systemSchemas: Set<String> {
         Set(PluginManager.shared.systemSchemaNames(for: databaseType))
     }
 
@@ -84,7 +86,8 @@ final class DatabaseTreeOutlineCoordinator: NSObject {
         viewModel = view.viewModel
 
         let activeChanged = activeDatabase != view.activeDatabase || activeSchema != view.activeSchema
-        let changed = searchText != view.searchText
+        let changed = connectionChanged
+            || searchText != view.searchText
             || isConnected != view.isConnected
             || activeChanged
             || pendingTruncates != view.pendingTruncates
@@ -103,14 +106,12 @@ final class DatabaseTreeOutlineCoordinator: NSObject {
             persistActiveExpansion()
         }
 
-        if !hasRenderedOnce {
+        guard hasRenderedOnce, !changed else {
             hasRenderedOnce = true
             refresh()
-        } else if changed {
-            refresh()
-        } else {
-            syncSelectionToModel()
+            return
         }
+        syncSelectionToModel()
     }
 
     private func persistActiveExpansion() {
@@ -153,6 +154,11 @@ final class DatabaseTreeOutlineCoordinator: NSObject {
     private func snapshotDependencies() {
         _ = service.databaseListState(for: connectionId)
         _ = sidebarState?.recentTables
+        /// Both feed `rootNodes()`: the layout picks the root's shape and the filter picks which
+        /// databases survive into it. Left unobserved, hiding a database in the filter popover
+        /// changed nothing until some other edit happened to rebuild the tree.
+        _ = sidebarState?.sidebarLayout
+        _ = sidebarState?.databaseFilterSelected
         /// One token covers every table, routine and per-schema load for this connection, which is
         /// the whole reactive surface the flat and hierarchical shapes read.
         _ = schemaService.generationToken(for: connectionId)
@@ -187,6 +193,7 @@ final class DatabaseTreeOutlineCoordinator: NSObject {
         guard let outlineView else { return }
         isReloading = true
         childrenCache.removeAll()
+        invalidateRowConfiguration()
         outlineView.reloadData()
         applyDesiredExpansion()
         syncSelectionToModel()
@@ -198,13 +205,11 @@ final class DatabaseTreeOutlineCoordinator: NSObject {
     /// place. Reloading would throw away every hosted SwiftUI view to repaint one glyph.
     private func refreshVisibleRows() {
         guard let outlineView else { return }
-        let context = rowContext()
-        let actions = rowActions()
         for row in 0..<outlineView.numberOfRows {
             guard let node = outlineView.item(atRow: row) as? DatabaseTreeNode,
                   let cell = outlineView.view(atColumn: 0, row: row, makeIfNecessary: false) as? DatabaseTreeCellView
             else { continue }
-            cell.configure(node: node, context: context, actions: actions)
+            cell.configure(node: node, context: rowContext, actions: rowActions)
         }
     }
 
@@ -226,526 +231,6 @@ final class DatabaseTreeOutlineCoordinator: NSObject {
         FavoriteTablesStorage.shared.toggle(
             name: entry.name, schema: entry.schema, database: entry.database, connectionId: connectionId
         )
-    }
-
-    // MARK: - Node building
-
-    private func node(id: String, kind: DatabaseTreeNode.Kind) -> DatabaseTreeNode {
-        if let existing = nodeCache[id] {
-            existing.kind = kind
-            return existing
-        }
-        let created = DatabaseTreeNode(id: id, kind: kind)
-        nodeCache[id] = created
-        return created
-    }
-
-    private func resolvedChildren(of item: Any?) -> [DatabaseTreeNode] {
-        let key = (item as? DatabaseTreeNode)?.id ?? ""
-        if let cached = childrenCache[key] { return cached }
-        let built = buildChildren(of: item as? DatabaseTreeNode)
-        childrenCache[key] = built
-        return built
-    }
-
-    private func buildChildren(of node: DatabaseTreeNode?) -> [DatabaseTreeNode] {
-        guard let node else { return rootNodes() }
-        switch node.kind {
-        case .recentSection:
-            return recentTableRefs().map {
-                self.node(id: DatabaseTreeNode.recentTableId($0), kind: .recentTable($0))
-            }
-        case .database(let metadata):
-            return supportsSchemaLevel
-                ? schemaNodes(database: metadata.name)
-                : objectNodes(database: metadata.name, schema: nil)
-        case .schema(let database, let schema):
-            return objectNodes(database: database, schema: schema)
-        case .table(let ref):
-            return ref.table.type == .partitionedTable ? partitionNodes(of: ref) : []
-        case .objectKindSection(let kind):
-            return flatObjectNodes(for: kind)
-        case .hierarchicalSchemaSection(let schema):
-            return hierarchicalTableNodes(schema: schema)
-        case .redisKeysSection:
-            return redisChildren(of: nil)
-        case .redisNode(let redisNode):
-            return redisChildren(of: redisNode)
-        case .recentTable, .routine, .status:
-            return []
-        }
-    }
-
-    private func partitionNodes(of ref: DatabaseTreeTableRef) -> [DatabaseTreeNode] {
-        let parentId = DatabaseTreeNode.tableId(ref)
-        let state = service.partitionsLoadState(
-            connectionId: connectionId, database: ref.database, schema: ref.schema, table: ref.table.name
-        )
-        switch state {
-        case .idle, .loading:
-            return [statusNode(parentId: parentId, status: .loading)]
-        case .failed(let message):
-            return [statusNode(parentId: parentId, status: .error(message))]
-        case .loaded(let partitions):
-            if partitions.isEmpty { return [statusNode(parentId: parentId, status: .empty)] }
-            return partitions.map { partition in
-                let childRef = DatabaseTreeTableRef(database: ref.database, schema: ref.schema, table: partition)
-                return node(id: DatabaseTreeNode.tableId(childRef), kind: .table(childRef))
-            }
-        }
-    }
-
-    /// Which shape the root takes. The three sidebar modes used to be three views; they are one
-    /// outline now and this is the only thing that still differs between them.
-    private var rootShape: SidebarRootShape {
-        SidebarRootShapeResolver.resolve(
-            groupingStrategy: PluginManager.shared.databaseGroupingStrategy(for: databaseType),
-            sidebarLayout: sidebarState?.sidebarLayout ?? .flat,
-            supportsDatabaseTree: PluginManager.shared.supportsDatabaseTree(for: databaseType)
-        )
-    }
-
-    private func rootNodes() -> [DatabaseTreeNode] {
-        switch rootShape {
-        case .databaseTree: return databaseTreeRootNodes()
-        case .flat: return flatRootNodes()
-        case .hierarchicalSchema: return hierarchicalRootNodes()
-        }
-    }
-
-    private func databaseTreeRootNodes() -> [DatabaseTreeNode] {
-        var nodes: [DatabaseTreeNode] = []
-        if !recentTableRefs().isEmpty {
-            nodes.append(node(id: DatabaseTreeNode.recentSectionId, kind: .recentSection))
-        }
-        let visible = DatabaseTreeVisibility.visible(
-            databases: service.databases(for: connectionId),
-            selected: sidebarState?.databaseFilterSelected ?? [],
-            activeDatabase: mainCoordinator?.browseDatabaseName ?? activeDatabase
-        )
-        let matched = searchText.isEmpty ? visible : visible.filter { databaseMatchesSearch($0) }
-        var seen = Set<String>()
-        nodes += matched
-            .filter { seen.insert($0.id).inserted }
-            .map { node(id: DatabaseTreeNode.databaseId($0.name), kind: .database($0)) }
-        return nodes
-    }
-
-    private var browsingDatabase: String? {
-        let name = mainCoordinator?.browseDatabaseName ?? activeDatabase ?? ""
-        return name.isEmpty ? nil : name
-    }
-
-    private func flatRootNodes() -> [DatabaseTreeNode] {
-        var nodes: [DatabaseTreeNode] = []
-        if !recentTableRefs().isEmpty {
-            nodes.append(node(id: DatabaseTreeNode.recentSectionId, kind: .recentSection))
-        }
-        nodes += visibleObjectKinds().map {
-            node(id: DatabaseTreeNode.objectKindSectionId($0), kind: .objectKindSection($0))
-        }
-        if sidebarState?.redisKeyTreeViewModel != nil {
-            nodes.append(node(id: DatabaseTreeNode.redisKeysSectionId, kind: .redisKeysSection))
-        }
-        return nodes
-    }
-
-    /// The section list is the same rule the flat list used, so a kind that was hidden before stays
-    /// hidden: Tables always shows, anything else needs both the capability and something in it.
-    private func visibleObjectKinds() -> [SidebarObjectKind] {
-        guard let viewModel else { return [] }
-        let capabilities = viewModel.capabilities(for: connectionId)
-        return SidebarObjectKind.allCases.filter { kind in
-            viewModel.sectionShouldRender(
-                kind: kind,
-                itemCount: flatItemCount(for: kind),
-                capabilities: capabilities
-            )
-        }
-    }
-
-    private func flatItemCount(for kind: SidebarObjectKind) -> Int {
-        guard let viewModel else { return 0 }
-        if kind.isRoutine {
-            return viewModel.filteredRoutines(of: kind, from: schemaService.routines(for: connectionId)).count
-        }
-        return viewModel.filteredTables(of: kind, from: schemaService.tables(for: connectionId)).count
-    }
-
-    private func flatObjectNodes(for kind: SidebarObjectKind) -> [DatabaseTreeNode] {
-        guard let viewModel else { return [] }
-        let database = browsingDatabase ?? ""
-        if kind.isRoutine {
-            return viewModel.filteredRoutines(of: kind, from: schemaService.routines(for: connectionId))
-                .map { routine in
-                    let ref = DatabaseTreeRoutineRef(database: database, schema: routine.schema, routine: routine)
-                    return node(id: DatabaseTreeNode.routineId(ref), kind: .routine(ref))
-                }
-        }
-        return viewModel.filteredTables(of: kind, from: schemaService.tables(for: connectionId))
-            .map { table in
-                let ref = DatabaseTreeTableRef(database: database, schema: table.schema, table: table)
-                return node(id: DatabaseTreeNode.tableId(ref), kind: .table(ref))
-            }
-    }
-
-    private func hierarchicalRootNodes() -> [DatabaseTreeNode] {
-        var nodes: [DatabaseTreeNode] = []
-        if !recentTableRefs().isEmpty {
-            nodes.append(node(id: DatabaseTreeNode.recentSectionId, kind: .recentSection))
-        }
-        let hidden = systemSchemas
-        nodes += schemaService.schemas(for: connectionId)
-            .filter { !hidden.contains($0) }
-            .filter { searchText.isEmpty || hierarchicalSchemaMatches($0) }
-            .map {
-                node(id: DatabaseTreeNode.hierarchicalSchemaSectionId($0), kind: .hierarchicalSchemaSection(schema: $0))
-            }
-        return nodes
-    }
-
-    private func hierarchicalSchemaMatches(_ schema: String) -> Bool {
-        DatabaseTreeFilter.hierarchicalSchemaIsVisible(
-            schema,
-            searchText: searchText,
-            isLoaded: isSchemaLoaded(schema),
-            tables: schemaService.tables(for: connectionId, schema: schema)
-        )
-    }
-
-    private func isSchemaLoaded(_ schema: String) -> Bool {
-        if case .loaded = schemaService.schemaState(for: connectionId, schema: schema) { return true }
-        return false
-    }
-
-    private func hierarchicalTableNodes(schema: String) -> [DatabaseTreeNode] {
-        let parentId = DatabaseTreeNode.hierarchicalSchemaSectionId(schema)
-        switch schemaService.schemaState(for: connectionId, schema: schema) {
-        case .idle, .loading:
-            return [statusNode(parentId: parentId, status: .loading)]
-        case .failed(let message):
-            return [statusNode(parentId: parentId, status: .error(message))]
-        case .loaded:
-            let tables = DatabaseTreeFilter.hierarchicalTables(
-                schemaService.tables(for: connectionId, schema: schema), schema: schema, searchText: searchText
-            )
-            guard !tables.isEmpty else { return [statusNode(parentId: parentId, status: .empty)] }
-            let database = browsingDatabase ?? ""
-            return tables.map { table in
-                let ref = DatabaseTreeTableRef(database: database, schema: schema, table: table)
-                return node(id: DatabaseTreeNode.tableId(ref), kind: .table(ref))
-            }
-        }
-    }
-
-    private func redisChildren(of parent: RedisKeyNode?) -> [DatabaseTreeNode] {
-        guard let keyTree = sidebarState?.redisKeyTreeViewModel else { return [] }
-        if let parent {
-            guard case .namespace(_, _, let children, _) = parent else { return [] }
-            return children.map { node(id: DatabaseTreeNode.redisNodeId($0), kind: .redisNode($0)) }
-        }
-        if keyTree.isLoading {
-            return [statusNode(parentId: DatabaseTreeNode.redisKeysSectionId, status: .loading)]
-        }
-        let roots = keyTree.displayNodes(searchText: searchText)
-        guard !roots.isEmpty else {
-            return [statusNode(parentId: DatabaseTreeNode.redisKeysSectionId, status: .empty)]
-        }
-        var nodes = roots.map { node(id: DatabaseTreeNode.redisNodeId($0), kind: .redisNode($0)) }
-        if keyTree.isTruncated {
-            nodes.append(
-                statusNode(
-                    parentId: DatabaseTreeNode.redisKeysSectionId,
-                    status: .truncated(RedisKeyTreeTruncation.message(limit: RedisKeyTreeViewModel.maxKeys))
-                )
-            )
-        }
-        return nodes
-    }
-
-    private func recentTableRefs() -> [DatabaseTreeTableRef] {
-        guard let sidebarState, showRecentTables else { return [] }
-        let database = mainCoordinator?.browseDatabaseName ?? activeDatabase ?? ""
-        return sidebarState.recentEntries(inDatabase: database).compactMap { entry -> DatabaseTreeTableRef? in
-            if !searchText.isEmpty, !DatabaseTreeFilter.matches(searchText, entry.name) { return nil }
-            return DatabaseTreeTableRef(database: database, schema: entry.schema, table: entry.tableInfo)
-        }
-    }
-
-    private func schemaNodes(database: String) -> [DatabaseTreeNode] {
-        let parentId = DatabaseTreeNode.databaseId(database)
-        switch service.schemaListState(connectionId: connectionId, database: database) {
-        case .idle, .loading:
-            return [statusNode(parentId: parentId, status: .loading)]
-        case .failed(let message):
-            return [statusNode(parentId: parentId, status: .error(message))]
-        case .loaded(let schemas):
-            let visible = DatabaseTreeFilter.visibleSchemas(
-                schemas,
-                systemSchemas: systemSchemas,
-                searchText: searchText,
-                contentMatches: { schemaContentMatchesSearch(database: database, schema: $0) }
-            )
-            if visible.isEmpty { return [statusNode(parentId: parentId, status: .empty)] }
-            return visible.map {
-                node(id: DatabaseTreeNode.schemaId(database: database, schema: $0), kind: .schema(database: database, schema: $0))
-            }
-        }
-    }
-
-    private func objectNodes(database: String, schema: String?) -> [DatabaseTreeNode] {
-        let parentId = schema.map { DatabaseTreeNode.schemaId(database: database, schema: $0) }
-            ?? DatabaseTreeNode.databaseId(database)
-        switch service.tablesLoadState(connectionId: connectionId, database: database, schema: schema) {
-        case .idle, .loading:
-            return [statusNode(parentId: parentId, status: .loading)]
-        case .failed(let message):
-            return [statusNode(parentId: parentId, status: .error(message))]
-        case .loaded:
-            return loadedObjectNodes(database: database, schema: schema, parentId: parentId)
-        }
-    }
-
-    private func loadedObjectNodes(database: String, schema: String?, parentId: String) -> [DatabaseTreeNode] {
-        let tables = DatabaseTreeFilter.filteredTables(
-            service.tables(connectionId: connectionId, database: database, schema: schema), searchText: searchText
-        )
-        let routines = DatabaseTreeFilter.filteredRoutines(
-            service.routines(connectionId: connectionId, database: database, schema: schema), searchText: searchText
-        )
-        let routinesState = service.routinesLoadState(connectionId: connectionId, database: database, schema: schema)
-
-        guard !tables.isEmpty || !routines.isEmpty else {
-            switch routinesState {
-            case .failed(let message): return [statusNode(parentId: parentId, status: .error(message))]
-            case .loaded: return [statusNode(parentId: parentId, status: .empty)]
-            case .idle, .loading: return [statusNode(parentId: parentId, status: .loading)]
-            }
-        }
-
-        var nodes: [DatabaseTreeNode] = tables.map { table in
-            let ref = DatabaseTreeTableRef(database: database, schema: schema, table: table)
-            return node(id: DatabaseTreeNode.tableId(ref), kind: .table(ref))
-        }
-        nodes += routines.map { routine in
-            let ref = DatabaseTreeRoutineRef(database: database, schema: schema, routine: routine)
-            return node(id: DatabaseTreeNode.routineId(ref), kind: .routine(ref))
-        }
-        if case .failed(let message) = routinesState {
-            nodes.append(statusNode(parentId: parentId, status: .error(message)))
-        }
-        return nodes
-    }
-
-    private func statusNode(parentId: String, status: DatabaseTreeNode.Status) -> DatabaseTreeNode {
-        node(id: DatabaseTreeNode.statusId(parentId: parentId, status: status), kind: .status(status))
-    }
-
-    // MARK: - Search
-
-    private func databaseMatchesSearch(_ metadata: DatabaseMetadata) -> Bool {
-        if DatabaseTreeFilter.matches(searchText, metadata.name) { return true }
-        if case .loaded(let schemas) = service.schemaListState(connectionId: connectionId, database: metadata.name) {
-            if schemas.contains(where: { DatabaseTreeFilter.matches(searchText, $0) }) { return true }
-            for schema in schemas where schemaContentMatchesSearch(database: metadata.name, schema: schema) {
-                return true
-            }
-        }
-        return schemaContentMatchesSearch(database: metadata.name, schema: nil)
-    }
-
-    private func schemaContentMatchesSearch(database: String, schema: String?) -> Bool {
-        if let schema, DatabaseTreeFilter.matches(searchText, schema) { return true }
-        let tables = service.tables(connectionId: connectionId, database: database, schema: schema)
-        if tables.contains(where: { DatabaseTreeFilter.matches(searchText, $0.name) }) { return true }
-        let routines = service.routines(connectionId: connectionId, database: database, schema: schema)
-        return routines.contains { DatabaseTreeFilter.matches(searchText, $0.name) }
-    }
-
-    // MARK: - Expansion
-
-    private func applyDesiredExpansion() {
-        guard let outlineView else { return }
-        isApplyingExpansion = true
-        defer { isApplyingExpansion = false }
-        let searching = !searchText.isEmpty
-        for rootNode in resolvedChildren(of: nil) where rootNode.id == DatabaseTreeNode.recentSectionId {
-            setExpanded(rootNode, searching || (viewModel?.isRecentsExpanded ?? true))
-        }
-        for sectionNode in resolvedChildren(of: nil) {
-            switch sectionNode.kind {
-            case .objectKindSection(let kind):
-                let hasMatches = flatItemCount(for: kind) > 0
-                setExpanded(sectionNode, viewModel?.effectiveExpanded(kind: kind, hasMatches: hasMatches) ?? true)
-            case .redisKeysSection:
-                setExpanded(sectionNode, searching || (viewModel?.isRedisKeysExpanded ?? true))
-            case .hierarchicalSchemaSection(let schema):
-                let want = searching
-                    ? hierarchicalSchemaMatches(schema)
-                    : windowState?.expandedTreeSchemas.contains(schema) ?? false
-                setExpanded(sectionNode, want)
-                if outlineView.isItemExpanded(sectionNode) { triggerLoad(for: sectionNode) }
-            default:
-                break
-            }
-        }
-        for databaseNode in resolvedChildren(of: nil) {
-            guard case .database(let metadata) = databaseNode.kind else { continue }
-            let want = searching
-                ? databaseMatchesSearch(metadata)
-                : windowState?.expandedTreeDatabases.contains(metadata.name) ?? false
-            setExpanded(databaseNode, want)
-            guard outlineView.isItemExpanded(databaseNode) else { continue }
-            triggerLoad(for: databaseNode)
-            guard supportsSchemaLevel else {
-                restorePartitionExpansion(under: databaseNode)
-                continue
-            }
-            for schemaNode in resolvedChildren(of: databaseNode) {
-                guard case .schema(let database, let schema) = schemaNode.kind else { continue }
-                let wantSchema = searching
-                    ? DatabaseTreeFilter.matches(searchText, schema) || schemaContentMatchesSearch(database: database, schema: schema)
-                    : windowState?.expandedTreeDatabaseSchemas.contains(DatabaseSchemaKey(database: database, schema: schema)) ?? false
-                setExpanded(schemaNode, wantSchema)
-                if outlineView.isItemExpanded(schemaNode) {
-                    triggerLoad(for: schemaNode)
-                    restorePartitionExpansion(under: schemaNode)
-                }
-            }
-        }
-    }
-
-    private func restorePartitionExpansion(under parent: DatabaseTreeNode) {
-        guard searchText.isEmpty, let outlineView, let windowState else { return }
-        for tableNode in resolvedChildren(of: parent) {
-            guard case .table(let ref) = tableNode.kind, ref.table.type == .partitionedTable else { continue }
-            let key = DatabaseTableKey(database: ref.database, schema: ref.schema, table: ref.table.name)
-            guard windowState.expandedTreeTables.contains(key) else { continue }
-            setExpanded(tableNode, true)
-            guard outlineView.isItemExpanded(tableNode) else { continue }
-            triggerLoad(for: tableNode)
-            restorePartitionExpansion(under: tableNode)
-        }
-    }
-
-    private func setExpanded(_ node: DatabaseTreeNode, _ expanded: Bool) {
-        guard let outlineView else { return }
-        if expanded, !outlineView.isItemExpanded(node) {
-            outlineView.expandItem(node)
-        } else if !expanded, outlineView.isItemExpanded(node) {
-            outlineView.collapseItem(node)
-        }
-    }
-
-    private func recordExpansion(_ node: DatabaseTreeNode, expanded: Bool) {
-        switch node.kind {
-        case .recentSection:
-            viewModel?.isRecentsExpanded = expanded
-        case .objectKindSection(let kind):
-            viewModel?.expanded[kind] = expanded
-        case .redisKeysSection:
-            viewModel?.isRedisKeysExpanded = expanded
-        case .hierarchicalSchemaSection(let schema):
-            if expanded {
-                windowState?.expandedTreeSchemas.insert(schema)
-            } else {
-                windowState?.expandedTreeSchemas.remove(schema)
-            }
-        case .database(let metadata):
-            if expanded {
-                windowState?.expandedTreeDatabases.insert(metadata.name)
-            } else {
-                windowState?.expandedTreeDatabases.remove(metadata.name)
-            }
-        case .schema(let database, let schema):
-            let key = DatabaseSchemaKey(database: database, schema: schema)
-            if expanded {
-                windowState?.expandedTreeDatabaseSchemas.insert(key)
-            } else {
-                windowState?.expandedTreeDatabaseSchemas.remove(key)
-            }
-        case .table(let ref):
-            let key = DatabaseTableKey(database: ref.database, schema: ref.schema, table: ref.table.name)
-            if expanded {
-                windowState?.expandedTreeTables.insert(key)
-            } else {
-                windowState?.expandedTreeTables.remove(key)
-            }
-        case .recentTable, .routine, .status, .redisNode:
-            break
-        }
-    }
-
-    private func triggerLoad(for node: DatabaseTreeNode) {
-        switch node.kind {
-        case .database(let metadata):
-            if supportsSchemaLevel {
-                if isIdle(service.schemaListState(connectionId: connectionId, database: metadata.name)) {
-                    Task { await service.loadSchemas(connectionId: connectionId, database: metadata.name) }
-                }
-                loadExternalSchemaNames(database: metadata.name)
-            } else {
-                loadObjects(database: metadata.name, schema: nil)
-            }
-        case .schema(let database, let schema):
-            loadObjects(database: database, schema: schema)
-        case .table(let ref):
-            loadPartitions(ref)
-        case .hierarchicalSchemaSection(let schema):
-            loadHierarchicalSchemaTables(schema)
-        case .recentSection, .recentTable, .routine, .status,
-             .objectKindSection, .redisKeysSection, .redisNode:
-            break
-        }
-    }
-
-    private func loadHierarchicalSchemaTables(_ schema: String) {
-        guard case .idle = schemaService.schemaState(for: connectionId, schema: schema),
-              let driver = DatabaseManager.shared.driver(for: connectionId) else { return }
-        let connectionId = connectionId
-        Task { await schemaService.loadSchemaTables(connectionId: connectionId, schema: schema, driver: driver) }
-    }
-
-    private func loadExternalSchemaNames(database: String) {
-        guard let session = DatabaseManager.shared.session(for: connectionId),
-              DatabaseManager.shared.browseDatabaseName(for: session.connection) == database,
-              let driver = DatabaseManager.shared.driver(for: connectionId)
-        else { return }
-        let connectionId = connectionId
-        Task {
-            await ExternalSchemaTracker.shared.load(
-                connectionId: connectionId,
-                database: database,
-                driver: driver
-            )
-        }
-    }
-
-    private func loadPartitions(_ ref: DatabaseTreeTableRef) {
-        guard ref.table.type == .partitionedTable else { return }
-        let state = service.partitionsLoadState(
-            connectionId: connectionId, database: ref.database, schema: ref.schema, table: ref.table.name
-        )
-        guard isIdle(state) else { return }
-        Task {
-            await service.loadPartitions(
-                connectionId: connectionId, database: ref.database, schema: ref.schema, table: ref.table.name
-            )
-        }
-    }
-
-    private func loadObjects(database: String, schema: String?) {
-        if isIdle(service.tablesLoadState(connectionId: connectionId, database: database, schema: schema)) {
-            Task { await service.loadTables(connectionId: connectionId, database: database, schema: schema) }
-        }
-        if isIdle(service.routinesLoadState(connectionId: connectionId, database: database, schema: schema)) {
-            Task { await service.loadRoutines(connectionId: connectionId, database: database, schema: schema) }
-        }
-    }
-
-    private func isIdle<Value>(_ state: MetadataLoadState<Value>) -> Bool {
-        if case .idle = state { return true }
-        return false
     }
 
     // MARK: - Selection / open
@@ -876,7 +361,30 @@ final class DatabaseTreeOutlineCoordinator: NSObject {
         }
     }
 
-    private func rowContext() -> DatabaseTreeRowContext {
+    /// Every visible row is handed the same context and the same action set, and both are pure
+    /// functions of the inputs `update(from:)` already tracks, so they are built once per refresh
+    /// instead of once per row. Rebuilding them in `viewFor` allocated a fresh set of closures for
+    /// every row the outline drew, on every reload and every scroll.
+    private var rowContext: DatabaseTreeRowContext {
+        if let cachedRowContext { return cachedRowContext }
+        let context = makeRowContext()
+        cachedRowContext = context
+        return context
+    }
+
+    private var rowActions: DatabaseTreeRowActions {
+        if let cachedRowActions { return cachedRowActions }
+        let actions = makeRowActions()
+        cachedRowActions = actions
+        return actions
+    }
+
+    private func invalidateRowConfiguration() {
+        cachedRowContext = nil
+        cachedRowActions = nil
+    }
+
+    private func makeRowContext() -> DatabaseTreeRowContext {
         DatabaseTreeRowContext(
             databaseType: databaseType,
             activeDatabase: activeDatabase,
@@ -903,7 +411,7 @@ final class DatabaseTreeOutlineCoordinator: NSObject {
         )
     }
 
-    private func rowActions() -> DatabaseTreeRowActions {
+    private func makeRowActions() -> DatabaseTreeRowActions {
         DatabaseTreeRowActions(
             coordinator: mainCoordinator,
             isReadOnly: mainCoordinator?.safeModeLevel.blocksAllWrites ?? false,
@@ -1001,7 +509,7 @@ extension DatabaseTreeOutlineCoordinator: NSOutlineViewDelegate {
         guard let node = item as? DatabaseTreeNode else { return nil }
         let cell = outlineView.makeView(withIdentifier: Self.cellIdentifier, owner: self) as? DatabaseTreeCellView
             ?? makeCell()
-        cell.configure(node: node, context: rowContext(), actions: rowActions())
+        cell.configure(node: node, context: rowContext, actions: rowActions)
         return cell
     }
